Initial version of SYCL-BLAS library
This version allow user to build sycl_blas library and use it by
including sycl_blas.h, or directly use the headers only by including
sycl_blas.hpp. All the files have been broken in to .h and .hpp.
Those who had sycl kernel on it has been instantiated by .cpp file for the
library.

A template CPP using py_gen folder is used to auto generate the
instantiation of .cpp files.

Added the py_gen submodule.

Updated CMakefiles to build library on request.

+++ bench/clblast/clblast_benchmark_level2.cpp
@@ -54,8 +54,8 @@ BENCHMARK(gemv, clblast_level_2) {
       static_cast<size_t>(m) * static_cast<size_t>(n) * static_cast<size_t>(2);
 
   size_t lda = m;
-  long incX = 1;
-  long incY = 1;
+  int incX = 1;
+  int incY = 1;
 
   // Specify the layout. As with GEMM, this needs to be kColMajor, and results
   // in errors otherwise. It may be that this is incorrect (especially for
